deep_learning_marcanthia/requirements.txt
absl-py==0.6.1
albumentations==0.1.2
astor==0.7.1
backcall==0.1.0
bleach==3.0.2
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
Cython==0.29
dask==0.20.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
gast==0.2.0
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-bigquery==1.6.0
google-cloud-core==0.28.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
grpcio==1.16.0
h5py==2.8.0
idna==2.7
imageio==2.4.1
imgaug==0.2.6
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numpy==1.15.3
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkg-resources==0.0.0
progressbar2==3.38.0
prometheus-client==0.4.1
prompt-toolkit==2.0.5
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py-cpuinfo==4.0.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
rmate==1.0.3
rsa==4.0
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
tensorlayer==1.11.0
tensornets==0.3.5
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==0.4.1
torchsummary==1.5.1
torchvision==0.2.1
tornado==5.1.1
tqdm==4.27.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
